# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для телефонного справочника",
        "title": "Telephone Book API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/workers": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workers"
                ],
                "summary": "Обновить работника",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Институт",
                        "name": "institute",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Email работника",
                        "name": "email",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Новые данные работника",
                        "name": "worker",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/workers.UpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/workers.UpdateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workers"
                ],
                "summary": "Создать работника",
                "parameters": [
                    {
                        "description": "Данные работника",
                        "name": "worker",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/workers.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/workers.CreateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workers"
                ],
                "summary": "Удалить работника",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Институт",
                        "name": "institute",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Email работника",
                        "name": "email",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/workers.DeleteResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "response.Response": {
            "description": "Базовый ответ API",
            "type": "object",
            "properties": {
                "error": {
                    "description": "Сообщение об ошибке, если есть. Пример: \"invalid request\"",
                    "type": "string"
                },
                "status": {
                    "description": "Статус ответа: Ok или Error. Пример: \"Ok\"",
                    "type": "string"
                }
            }
        },
...
        "workers.CreateRequest": {
            "type": "object",
            "required": [
                "email",
                "institute",
                "name",
                "phone_number",
                "surname"
            ],
            "properties": {
                "birth_date": {
                    "type": "string"
                },
                "cabinet": {
                    "type": "string"
                },
                "department": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "institute": {
                    "type": "string"
                },
                "middle_name": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "position": {
                    "type": "string"
                },
                "section": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        },
        "workers.CreateResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Сообщение об ошибке, если есть. Пример: \"invalid request\"",
                    "type": "string"
                },
                "status": {
                    "description": "Статус ответа: Ok или Error. Пример: \"Ok\"",
                    "type": "string"
                },
                "user_id": {
                    "description": "ID созданного работника",
                    "type": "integer"
                }
            }
        },
        "workers.DeleteResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Сообщение об ошибке, если есть. Пример: \"invalid request\"",
                    "type": "string"
                },
                "status": {
                    "description": "Статус ответа: Ok или Error. Пример: \"Ok\"",
                    "type": "string"
                }
            }
        },
...
        "workers.UpdateRequest": {
            "type": "object",
            "required": [
                "cabinet",
                "department",
                "email",
                "institute",
                "name",
                "phone_number",
                "position",
                "surname"
            ],
            "properties": {
                "birth_date": {
                    "type": "string"
                },
                "cabinet": {
                    "type": "string"
                },
                "department": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "institute": {
                    "type": "string"
                },
                "middle_name": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "photo": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "position": {
                    "type": "string"
                },
                "section": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        },
        "workers.UpdateResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Сообщение об ошибке, если есть. Пример: \"invalid request\"",
                    "type": "string"
                },
                "status": {
                    "description": "Статус ответа: Ok или Error. Пример: \"Ok\"",
                    "type": "string"
                }
            }
        }
    }
}